+++ src/utils/regexps.mjs
@@ -2,6 +2,6 @@
 
 export const eventListenerRegExp = /@([\w-]+)=['"]?$/
 export const dataEventIdJSRegExp = /^eventId[\d]{1,}$/
-export const customElementRegExp = /<([\w-]+)(:[\w]+)?=$/
+export const customElementRegExp = /<([\w-]+)\s*(:[\w]+)?\s*=$/
 export const styleAttributeRegExp = /style=['"]?[\w\s:;()-]*$/
 export const booleanAttributeRegExp = /\?([\w-]+)=['"]?$/